// src/app/page/master/asal/asal.interface.ts
import { EventEmitter } from '@angular/core';

export interface Asal {
	kdAsalHead?: number;
	kdDepartemen?: string;
	kode?: any;
	kodeExternal?: string;
	namaAsal?: string;
	namaAsalHead?: any;
	namaExternal?: string;
	reportDisplay?: string;
	statusEnabled?: boolean;
	id?: any;
	noRec?: string;
	label?: any;
	value?: any;
	id_kode?: any;
}